gimnasio_project/requirements.txt
asgiref==3.10.0
cffi==2.0.0
cryptography==46.0.3
Django==5.2.7
MarkupSafe==3.0.3
pillow==12.0.0
pycparser==2.23
python-http-client==3.3.7
sendgrid==6.12.5
sqlparse==0.5.3
Werkzeug==3.1.3
requests==2.32.3

# Dependencias adicionales para Docker/Producción
djangorestframework==3.14.0
drf-spectacular==0.27.0
reportlab==4.0.7
gunicorn==21.2.0
psycopg2-binary==2.9.10
python-decouple==3.8
whitenoise==6.6.0
